refactor: move nat64 and forward modules
