+++ Cargo.toml
@@ -6,10 +6,10 @@ members = [
     "cli/core",
     "cli/modules/common",
     "cli/modules/inspect",
-    "cli/modules/forward",
-    "cli/modules/pipeline",
-    "cli/modules/nat64",
+    "cli/modules/pipeline",    
     "modules/decap/cli",
+    "modules/forward/cli",
+    "modules/nat64/cli",    
     "modules/route/cli/neighbour",
     "modules/route/cli/route",
     "web",
