+++ agent/bootstrap/main.go
@@ -1,12 +1,13 @@
 package main
 
 //#cgo CFLAGS: -I../../ -I../../lib
-//#cgo LDFLAGS: -L../../build/modules/forward/ -lforward_cp
+//#cgo LDFLAGS: -L../../build/modules/forward/api -lforward_cp
 //#cgo LDFLAGS: -L../../build/lib/controlplane/agent -lagent
 //#cgo LDFLAGS: -L../../build/lib/controlplane/config -lconfig_cp
 //#cgo LDFLAGS: -L../../build/lib/dataplane/config -lconfig_dp
+//
 //#include "api/agent.h"
-//#include "modules/forward/controlplane.h"
+//#include "modules/forward/api/controlplane.h"
 //
 import "C"
