+++ controlplane/yncp/cfg.go
@@ -8,9 +8,9 @@ import (
 	"gopkg.in/yaml.v3"
 
 	"github.com/yanet-platform/yanet2/controlplane/internal/gateway"
-	"github.com/yanet-platform/yanet2/controlplane/modules/forward"
-	"github.com/yanet-platform/yanet2/controlplane/modules/nat64"
 	decap "github.com/yanet-platform/yanet2/modules/decap/controlplane"
+	forward "github.com/yanet-platform/yanet2/modules/forward/controlplane"
+	nat64 "github.com/yanet-platform/yanet2/modules/nat64/controlplane"
 	route "github.com/yanet-platform/yanet2/modules/route/controlplane"
 )
