+++ modules/nat64/controlplane/ffi.go
@@ -1,11 +1,11 @@
 package nat64
 
 //#cgo CFLAGS: -I../../../ -I../../../lib
-//#cgo LDFLAGS: -L../../../build/modules/nat64/ -lnat64_cp -llogging
+//#cgo LDFLAGS: -L../../../build/modules/nat64/api -lnat64_cp -llogging
 //#cgo LDFLAGS: -L../../../build/lib/logging/ -llogging
 //
 //#include "api/agent.h"
-//#include "modules/nat64/nat64cp.h"
+//#include "modules/nat64/api/nat64cp.h"
 import "C"
 
 import (
