+++ modules/nat64/tests/dataplane/nat64_test.c
@@ -27,12 +27,12 @@
 #include <rte_udp.h>
 
 /* Project headers */
+#include "api/nat64cp.h"
 #include "common.h"
 #include "dataplane/dpdk.h"
 #include "dataplane/module/module.h"
+#include "dataplane/nat64dp.h"
 #include "logging/log.h"
-#include "nat64cp.h"
-#include "nat64dp.h"
 #include "test.h"
 
 #ifdef DEBUG_NAT64
