+++ modules/nat64/tests/meson.build
@@ -1,0 +1,1 @@
+subdir('dataplane')
